Rework firewall exclusion rules
