+++ talpid-core/src/firewall/mod.rs
@@ -230,7 +230,7 @@ pub struct FirewallArguments {
     #[cfg(target_os = "macos")]
     /// This argument is required on macOS to know which group's traffic should be excluded, if at
     /// all.
-    pub exclusion_gid: Option<u32>,
+    pub exclusion_gid: u32,
 }
 
 /// State to enter during firewall init.
